rubocopによる検査

+++ app/controllers/products_controller.rb
@@ -36,8 +36,12 @@ class ProductsController < ApplicationController
 
   def destroy
     product = Product.find(params[:id])
-    product.destroy
-    redirect_to root_path
+    if user_signed_in? && current_user.id == product.user_id
+      product.destroy
+      redirect_to root_path
+    else
+      render :index
+    end
   end
 
   private
